Renamed ngTableParams to NgTableParams
When using the ngTable into our application we've encountered a jshint
error because the ngTableParams is a method that is starting with a
lower case instead of uppercase letter.

This issue can be sorted out by activating newcap in jshint so you can
ignore it.
/* jshint newcap:true */

However, I've renamed the ngTableParams with NgTableParams so we could
use jshint with default options and have no errors when placing ngTable
inside our app.